refactor(contracts): define the electronAPI surface once
apps/web/src/types/electron.d.ts hand-mirrored the preload API in 471 lines,
making 8 of the repo's top-20 duplication clones and leaving the two sides free
to drift silently. The namespace interfaces now live in @shiranami/contracts;
the preload modules implement them and the renderer d.ts references them.

Both ElectronAPI declarations now extend SharedElectronApi, so a namespace added
to the contract but never wired into the preload fails to compile at the
electronAPI object literal instead of quietly entering the renderer's surface.

Also types the db preload namespace properly: 21 Promise<unknown> / unknown[]
returns become real domain types, so the derive propagates safety instead of
propagating unknown. Type-only change; no channel, argument or payload moved.
