+++ apps/desktop/src/main/preload/api/app.ts
@@ -1,13 +1,9 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS } from '@shiranami/contracts';
+import { IPC_CHANNELS, type AppApi } from '@shiranami/contracts';
 
 const C = IPC_CHANNELS.app;
 
-export interface AppApi {
-  getVersion: () => Promise<string>;
-  openLogsFolder: () => Promise<void>;
-  getLocaleCountry: () => Promise<string>;
-}
+export type { AppApi };
 
 export const appApi: AppApi = {
   getVersion: () => invoke(C.getVersion),
