+++ apps/desktop/src/main/preload/api/debug.ts
@@ -1,17 +1,10 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS, type MainMetricsSnapshot } from '@shiranami/contracts';
+import { IPC_CHANNELS, type DebugApi, type MainMetricsSnapshot } from '@shiranami/contracts';
 import { createIpcListener } from '../ipc-listener';
 
 const C = IPC_CHANNELS.debug;
 
-export interface DebugApi {
-  /** Begin main-process metrics sampling (~1 Hz). Idempotent. */
-  start: () => Promise<void>;
-  /** Stop sampling and clear the interval. Idempotent. */
-  stop: () => Promise<void>;
-  /** Subscribe to main-process metric snapshots. Returns an unsubscribe fn. */
-  onMetrics: (callback: (snapshot: MainMetricsSnapshot) => void) => () => void;
-}
+export type { DebugApi };
 
 export const debugApi: DebugApi = {
   start: () => invoke(C.start),
