+++ apps/desktop/src/main/preload/api/lyrics.ts
@@ -1,17 +1,9 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS, type LyricsResult } from '@shiranami/contracts';
+import { IPC_CHANNELS, type LyricsApi } from '@shiranami/contracts';
 
 const C = IPC_CHANNELS.lyrics;
 
-export interface LyricsApi {
-  fetch: (
-    title: string,
-    artist: string,
-    album?: string,
-    duration?: number,
-    filePath?: string
-  ) => Promise<LyricsResult>;
-}
+export type { LyricsApi };
 
 export const lyricsApi: LyricsApi = {
   fetch: (title, artist, album, duration, filePath) =>
