+++ apps/desktop/src/main/preload/api/media.ts
@@ -1,22 +1,10 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS } from '@shiranami/contracts';
+import { IPC_CHANNELS, type MediaApi } from '@shiranami/contracts';
 import { createIpcListener } from '../ipc-listener';
 
 const C = IPC_CHANNELS.media;
 
-export interface MediaApi {
-  onCommand: (callback: (command: string) => void) => () => void;
-  sendPlaybackState: (state: {
-    isPlaying: boolean;
-    title: string;
-    artist: string;
-    album: string;
-    duration: number;
-    currentTime: number;
-    albumArt: string | null;
-  }) => Promise<void>;
-  clearState: () => Promise<void>;
-}
+export type { MediaApi };
 
 export const mediaApi: MediaApi = {
   onCommand: createIpcListener<string>(C.command),
