+++ apps/desktop/src/main/preload/api/playlist.ts
@@ -1,41 +1,18 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS } from '@shiranami/contracts';
+import {
+  IPC_CHANNELS,
+  type PlaylistApi,
+  type PlaylistExtractProgress,
+  type PlaylistExtractResult,
+} from '@shiranami/contracts';
 import { createIpcListener } from '../ipc-listener';
 
 const C = IPC_CHANNELS.playlist;
 
-interface PlaylistTrack {
-  id: string;
-  title: string;
-  uploader: string;
-  duration: number;
-  thumbnail: string;
-  url: string;
-  webpage_url: string;
-  /** 0..1 Spotify-match score; only present on the Spotify playlist path. */
-  matchConfidence?: number;
-  /** 'low' when the best YouTube candidate scored below the threshold. */
-  matchFlag?: 'low' | 'ok';
-}
-
-export interface PlaylistExtractResult {
-  /** Source playlist title, when the provider exposed one. */
-  title: string | null;
-  tracks: PlaylistTrack[];
-}
-
-export interface PlaylistApi {
-  extract: (url: string) => Promise<PlaylistExtractResult>;
-  cancel: () => Promise<void>;
-  onExtractProgress: (
-    callback: (data: { current: number; total: number; trackName: string }) => void
-  ) => () => void;
-}
+export type { PlaylistApi, PlaylistExtractResult };
 
 export const playlistApi: PlaylistApi = {
   extract: url => invoke(C.extract, url),
   cancel: () => invoke(C.cancel),
-  onExtractProgress: createIpcListener<{ current: number; total: number; trackName: string }>(
-    C.extractProgress
-  ),
+  onExtractProgress: createIpcListener<PlaylistExtractProgress>(C.extractProgress),
 };
