+++ apps/desktop/src/main/preload/api/radio.ts
@@ -1,18 +1,9 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS, type RadioStationInput, type RadioFavorite } from '@shiranami/contracts';
+import { IPC_CHANNELS, type RadioApi, type RadioFavoritesApi } from '@shiranami/contracts';
 
 const C = IPC_CHANNELS.radio.favorites;
 
-export interface RadioFavoritesApi {
-  getAll: () => Promise<RadioFavorite[]>;
-  add: (station: RadioStationInput) => Promise<RadioFavorite>;
-  remove: (stationUuid: string) => Promise<void>;
-  isFavorite: (stationUuid: string) => Promise<boolean>;
-}
-
-export interface RadioApi {
-  favorites: RadioFavoritesApi;
-}
+export type { RadioApi, RadioFavoritesApi };
 
 export const radioApi: RadioApi = {
   favorites: {
