+++ apps/desktop/src/main/preload/api/share.ts
@@ -1,24 +1,10 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS, type ShareImportResponse } from '@shiranami/contracts';
+import { IPC_CHANNELS, type ShareApi } from '@shiranami/contracts';
 import { createIpcListener } from '../ipc-listener';
 
 const C = IPC_CHANNELS.share;
 
-interface ShareCode {
-  code: string;
-  url: string;
-  expiresAt: string;
-}
-
-export interface ShareApi {
-  track: (trackId: string) => Promise<ShareCode>;
-  playlist: (playlistId: string) => Promise<ShareCode>;
-  // The main-process handler validates this against shareImportResponseSchema,
-  // so the renderer receives a fully-typed discriminated union, not raw unknown.
-  import: (code: string) => Promise<ShareImportResponse>;
-  cacheYoutubeId: (trackId: string, youtubeId: string) => Promise<void>;
-  onDeepLink: (callback: (code: string) => void) => () => void;
-}
+export type { ShareApi };
 
 export const shareApi: ShareApi = {
   track: trackId => invoke(C.track, trackId),
