+++ apps/desktop/src/main/preload/api/storage.ts
@@ -1,15 +1,9 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS, type DiskUsageResult } from '@shiranami/contracts';
+import { IPC_CHANNELS, type StorageApi } from '@shiranami/contracts';
 
 const C = IPC_CHANNELS.storage;
 
-export interface StorageApi {
-  /**
-   * Compute disk usage for the given watched library-folder paths. Returns one
-   * entry per physical volume the folders live on.
-   */
-  getUsage: (folderPaths: string[]) => Promise<DiskUsageResult>;
-}
+export type { StorageApi };
 
 export const storageApi: StorageApi = {
   getUsage: folderPaths => invoke(C.getUsage, folderPaths),
