+++ apps/desktop/src/main/preload/api/system.ts
@@ -1,12 +1,10 @@
-import { IPC_CHANNELS } from '@shiranami/contracts';
+import { IPC_CHANNELS, type SystemApi } from '@shiranami/contracts';
 import type { SystemNotice } from '@shiranami/contracts';
 import { createIpcListener } from '../ipc-listener';
 
 const C = IPC_CHANNELS.system;
 
-export interface SystemApi {
-  onNotice: (callback: (notice: SystemNotice) => void) => () => void;
-}
+export type { SystemApi };
 
 export const systemApi: SystemApi = {
   onNotice: createIpcListener<SystemNotice>(C.notice),
