+++ apps/desktop/src/main/preload/api/weather.ts
@@ -1,12 +1,9 @@
 import { invoke } from '../context-bridge';
-import { IPC_CHANNELS, type GeocodeResult, type WeatherCurrent } from '@shiranami/contracts';
+import { IPC_CHANNELS, type WeatherApi } from '@shiranami/contracts';
 
 const C = IPC_CHANNELS.weather;
 
-export interface WeatherApi {
-  geocode: (query: string) => Promise<GeocodeResult | null>;
-  getCurrent: (coords: { lat: number; lon: number }) => Promise<WeatherCurrent>;
-}
+export type { WeatherApi };
 
 export const weatherApi: WeatherApi = {
   geocode: query => invoke(C.geocode, query),
