+++ packages/contracts/src/index.ts
@@ -10,6 +10,8 @@ export * from './domain/smart-playlist';
 export * from './domain/download-queue';
 export * from './domain/dependencies';
 export * from './domain/radio';
+export * from './domain/playlist';
+export * from './domain/folder';
 export * from './ipc/channels';
 export * from './ipc/debug';
 export * from './ipc/metadata';
@@ -20,4 +22,5 @@ export * from './ipc/error-codes';
 export * from './ipc/history';
 export * from './ipc/system';
 export * from './ipc/storage';
+export * from './ipc/preload-api';
 export * from './share/dto';
